Modification of preferences code.

+++ bundles/org.eclipse.core.runtime/src/org/eclipse/core/runtime/Preferences.java
@@ -299,6 +299,10 @@ public class Preferences {
 	 * @see #validatePreferenceVersions
 	 */
 	public static void exportPreferences(IPath path) throws CoreException {
+		if (InternalPlatform.USE_LEGACY_PREFS) {
+			exporterInvoker("exportPreferences", path); //$NON-NLS-1$
+			return;
+		}
 		File file = path.toFile();
 		if (file.exists())
 			file.delete();
@@ -361,6 +365,10 @@ public class Preferences {
 	 * @see #validatePreferenceVersions
 	 */
 	public static void importPreferences(IPath path) throws CoreException {
+		if (InternalPlatform.USE_LEGACY_PREFS) {
+			exporterInvoker("importPreferences", path); //$NON-NLS-1$
+			return;
+		}
 		if (!path.toFile().exists()) {
 			String msg = Policy.bind("preferences.fileNotFound", path.toOSString()); //$NON-NLS-1$
 			throw new CoreException(new Status(IStatus.ERROR, Platform.PI_RUNTIME, 1, msg, null));
@@ -407,7 +415,7 @@ public class Preferences {
 		try {
 			return (IStatus) exporterInvoker("validatePreferenceVersions", file);
 		} catch (CoreException e) {
-			return null;
+			return e.getStatus();
 		}
 	}
